Add database set-up
